master:worker use case
After #31 three node cluster, where all nodes will have the master and
worker(or node for vanilla clusters) labels are just getting the
nfd-master daemonset deployed. Since they have the master label, the NFD
workers will not be scheduled.

Discussion started in a downstream distribution of NFD:
https://github.com/openshift/cluster-nfd-operator/issues/109

This Patch fix that, by adding support for master:worker type of nodes,
modifying the nodeAffinity rules on the nfd-worker daemonSet to allow any node with the "node" label
(independent if it is a master also) to get scheduled

Signed-off-by: Carlos Eduardo Arango Gutierrez <carangog@redhat.com>