docs: show full version number in sidebar
Change the sidebar customization so that the full version (e.g. v0.2.0)
is displayed, instead of the truncated "releae branch version" (e.g.
v0.2) in the top left corner of the page. The items in the version menu
are unchanged and will still show the shorter form.

Cherry picked from node-feature-discovery repo commit
4d19e1ab85b6cf4c6a2f6e556bf04654e4fb238e.