Use apimachinery error library to check for missing nfd instance
The existing error checking library in the cri-api package was used
to check for the case where the NFD instance is deleted. However,
this error checking library is used for GRPC specific errors, not
for errors related to the 'metav1' package. The 'metav1' package is
used by various k8s resources (e.g., DaemonSet, Service, etc.),
and thus, we should be checking for 'metav1' problems, not GRPC
problems.