Replace all logging methods with klog, if they do not use klog
Create a 'NodeFeatureDiscoveryLogger' object that is a wrapper
around klog itself. By using a wrapper and only using the new
'r.Log.*", we can ensure that the logging libraries used are
consistent across all files, and we can also ensure that it is easy
to change the logging library from klog to another logging library,
if that is desired in the future.

Signed-off-by: Carlos Eduardo Arango Gutierrez <carangog@redhat.com>
Co-Authored-by: Courtney Pacheco <cpacheco@redhat.com>