dockerfile: builder image configurable through a Makefile variable.
Switch over to the "non-point-release" version of the image. Now we
always use the latest patch version of golang with latest security
fixes, for example, without the need to manually bump the version after
every point release.

This patch also makes the builder image configurable through a Makefile
variable pointing to the go.mod go version, that way the go.mod file
becomes the source of truth for the go version.